# docs/swagger.yaml
definitions:
...
  entities.PindahEntityModel:
    properties:
      alamat:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      golongan_darah:
        type: string
      id:
        type: string
      jenis_kelamin:
        type: string
      modified_at:
        type: string
      modified_by:
        type: string
      nama_jemaat:
        type: string
      nomor_hp:
        type: string
      tanggal_lahir:
        type: string
      tempat_lahir:
        type: string
    type: object
...
  pindah.CreateRequest:
    properties:
      alamat:
        type: string
      gereja:
        type: string
      golongan_darah:
        type: string
      jenis_kelamin:
        type: string
      nama_jemaat:
        type: string
      nomor_hp:
        type: string
      tanggal_lahir:
        type: string
      tempat_lahir:
        type: string
    type: object
  pindah.CreateResponse:
    properties:
      alamat:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      golongan_darah:
        type: string
      id:
        type: string
      jenis_kelamin:
        type: string
      modified_at:
        type: string
      modified_by:
        type: string
      nama_jemaat:
        type: string
      nomor_hp:
        type: string
      tanggal_lahir:
        type: string
      tempat_lahir:
        type: string
    type: object
  pindah.CreateResponseDoc:
    properties:
      body:
        properties:
          data:
            $ref: '#/definitions/pindah.CreateResponse'
          meta:
            $ref: '#/definitions/response.Meta'
        type: object
    type: object
  pindah.DeleteResponse:
    properties:
      id:
        type: string
    type: object
  pindah.DeleteResponseDoc:
    properties:
      body:
        properties:
          data:
            $ref: '#/definitions/pindah.DeleteResponse'
          meta:
            $ref: '#/definitions/response.Meta'
        type: object
    type: object
  pindah.GetAllResponse:
    properties:
      datas:
        items:
          $ref: '#/definitions/entities.PindahEntityModel'
        type: array
    type: object
  pindah.GetAllResponseDoc:
    properties:
      body:
        properties:
          data:
            $ref: '#/definitions/pindah.GetAllResponse'
          meta:
            $ref: '#/definitions/response.Meta'
        type: object
    type: object
  pindah.GetByIDResponse:
    properties:
      alamat:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      golongan_darah:
        type: string
      id:
        type: string
      jenis_kelamin:
        type: string
      modified_at:
        type: string
      modified_by:
        type: string
      nama_jemaat:
        type: string
      nomor_hp:
        type: string
      tanggal_lahir:
        type: string
      tempat_lahir:
        type: string
    type: object
  pindah.GetByIDResponseDoc:
    properties:
      body:
        properties:
          data:
            $ref: '#/definitions/pindah.GetByIDResponse'
          meta:
            $ref: '#/definitions/response.Meta'
        type: object
    type: object
  pindah.UpdateRequest:
    properties:
      id:
        type: string
    required:
    - id
    type: object
  pindah.UpdateResponse:
    properties:
      alamat:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      golongan_darah:
        type: string
      id:
        type: string
      jenis_kelamin:
        type: string
      modified_at:
        type: string
      modified_by:
        type: string
      nama_jemaat:
        type: string
      nomor_hp:
        type: string
      tanggal_lahir:
        type: string
      tempat_lahir:
        type: string
    type: object
  pindah.UpdateResponseDoc:
    properties:
      body:
        properties:
          data:
            $ref: '#/definitions/pindah.UpdateResponse'
          meta:
            $ref: '#/definitions/response.Meta'
        type: object
    type: object
  response.Meta:
    properties:
      message:
        default: "true"
        type: string
      success:
        default: true
        type: boolean
    type: object
  response.errorResponse:
    properties:
      error:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
    type: object
info:
  contact: {}
paths:
...
  /pindah:
    get:
      consumes:
      - application/json
      description: Get ALl Pindah
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pindah.GetAllResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.errorResponse'
      summary: Get All Pindah
      tags:
      - Pindah
    patch:
      consumes:
      - application/json
      description: Update Pindah
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pindah.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pindah.UpdateResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.errorResponse'
      summary: Update Pindah
      tags:
      - Pindah
    post:
      consumes:
      - application/json
      description: Create Pindah
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pindah.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pindah.CreateResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.errorResponse'
      summary: Create Pindah
      tags:
      - Pindah
  /pindah/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Pindah
      parameters:
      - description: id path
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pindah.DeleteResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.errorResponse'
      summary: Delete Pindah
      tags:
      - Pindah
    get:
      consumes:
      - application/json
      description: Get Pindah By ID
      parameters:
      - description: id path
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pindah.GetByIDResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.errorResponse'
      summary: Get Pindah By ID
      tags:
      - Pindah
swagger: "2.0"